imotiko_backend/requirements.txt
Django>=4.2.0
pillow>=10.0.0
python-decouple>=3.8